requirements.txt
streamlit==1.32.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
requests==2.31.0
langdetect==1.0.9
groq==0.4.0
python-magic==0.4.27
validators==0.22.0